+++ CHANGES.rst
@@ -27,6 +27,10 @@ Changelog
   [gbastien]
 - Avoid add/edit advice popup closing when using WSC.
   [gbastien]
+- When sending email notifications, prepend `MeetingConfig.configGroup`
+  `full_label` when used and several `MeetingConfigs` have same title
+  so we know from which `MeetingConfig` the notification is sent.
+  [gbastien]
 
 4.2.19 (2025-04-02)
 -------------------